backend/requirements.txt
Django==4.2.19
djangorestframework==3.15.2
django-allauth==65.4.1
dj-rest-auth[with_social]==7.0.1
psycopg2==2.9.10
drf-spectacular==0.28.0
django-cors-headers==4.7.0
django-phonenumber-field[phonenumbers]==8.0.0
django-model-utils==5.0.0
pillow==11.1.0
django-money==3.5.3
django-extensions==3.2.3
faker
ruff
pre_commit==4.1.0
typing_extensions==4.12.2
django-filter==25.1
